correct Makefile for graphs, refine tests, change Graph to c++
* The Makefile was missing depends for the graphs
* The tests for BipartiteGraph were incomplete
* The graph functions have been rewritten to allow overriding them
  in the subclasses

+++ src/tests/graphtest.cpp
@@ -84,12 +84,11 @@ void GraphTest::parseFile(
 
    int input;
 
-   for( int i = 0; i < array.size(); ++i )
+   for( size_t i = 0; i < array.size(); ++i )
    {
       ASSERT_FALSE(stream.eof());
       ASSERT_TRUE(stream >> input);
       ASSERT_EQ(array[i], input);
    }
    stream.close();
-
 }
